Store and reuse overlays that have been deleted
When it deletes the overlays during refreshing, it will now move the
deleted overlays to a buffer local list. When creating new overlays, it
will first see if it there are any overlays in the list that can be
reused by moving them. Otherwise it will create a new one.
Linum does it this way. If there is any difference in performance, I
can't tell, but this solves a long-lasting issue with org-mode lists.

Fixes #6 (finally!)