Add function for integrating with evil-mode
Also some updates to docs based on checkdoc.
I still have to decide whether this is the way I want the evil
integration to work. Maybe there should be more safety for troubles with
packages not being loaded when the integration function is called.
Also need to test if the integration works with the optional arg.